Other: Refactor exports fo downcast converters helper module.
