+++ src/tableediting.js
@@ -10,7 +10,7 @@
 import Plugin from '@ckeditor/ckeditor5-core/src/plugin';
 import { upcastElementToElement } from '@ckeditor/ckeditor5-engine/src/conversion/upcast-converters';
 import upcastTable from './converters/upcasttable';
-import downcastTable, { downcastInsertCell, downcastInsertRow } from './converters/downcasttable';
+import { downcastInsertCell, downcastInsertRow, downcastInsertTable } from './converters/downcast';
 import InsertTableCommand from './inserttablecommand';
 import InsertRowCommand from './insertrowcommand';
 import InsertColumnCommand from './insertcolumncommand';
@@ -53,7 +53,7 @@ export default class TablesEditing extends Plugin {
 
 		// Table conversion.
 		conversion.for( 'upcast' ).add( upcastTable() );
-		conversion.for( 'downcast' ).add( downcastTable() );
+		conversion.for( 'downcast' ).add( downcastInsertTable() );
 
 		// Insert conversion
 		conversion.for( 'downcast' ).add( downcastInsertRow() );
